LPS-83417 create representor on a separate helper
